Test on Django 4.0 and Python 3.10
Add Django 4.0 on Python 3.10 to the test matrix.

Set `USE_TZ = True` in test settings since without it Django 4.0 raises a deprecation warning.

Add Python flags to fail for resource and deprecation warnings - this is something I like to do for libraries to check that warnings aren't passed on to users.